cms-bridge/requirements.txt
Flask==2.1.2
voluptuous>=0.12,<1
python-slugify==6.1.2
